fix: ナビゲーションバーの表示をScaffold内で作成できるように修正する
