+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-
-import 'widgets/custom_tab_bar.dart';
+import 'package:happyo/common/routes.dart';
+import 'package:happyo/page/home_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -16,43 +16,12 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const MyHomePage(),
-    );
-  }
-}
-
-class MyHomePage extends StatelessWidget {
-  const MyHomePage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: CustomTabBar(
-        tab: [
-          Tab(text: 'ネットワーク'),
-          Tab(text: 'おすすめ'),
-          Tab(text: 'セキュリティ'),
-          Tab(text: 'AI'),
-          Tab(text: 'AI2'),
-        ],
-        list: [
-          Center(
-            child: Text('ネットワーク', style: TextStyle(fontSize: 32.0)),
-          ),
-          Center(
-            child: Text('おすすめ', style: TextStyle(fontSize: 32.0)),
-          ),
-          Center(
-            child: Text('セキュリティ', style: TextStyle(fontSize: 32.0)),
-          ),
-          Center(
-            child: Text('AI', style: TextStyle(fontSize: 32.0)),
-          ),
-          Center(
-            child: Text('AI2', style: TextStyle(fontSize: 32.0)),
-          ),
-        ],
-      ),
+      initialRoute: Routes.index,
+      routes: {
+        Routes.index: (context) {
+          return HomePage();
+        },
+      },
     );
   }
 }
