+++ lib/page/home_page.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/material.dart';
+import 'package:happyo/widgets/custom_tab_bar.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('HAPPYO'),
+        centerTitle: false,
+      ),
+      drawer: Drawer(),
+      body: CustomTabBar(
+        tabs: [
+          Tab(text: 'ネットワーク'),
+          Tab(text: 'おすすめ'),
+          Tab(text: 'セキュリティ'),
+          Tab(text: 'AI'),
+          Tab(text: 'AI2'),
+        ],
+        list: [
+          Center(
+            child: Text('ネットワーク', style: TextStyle(fontSize: 32.0)),
+          ),
+          Center(
+            child: Text('おすすめ', style: TextStyle(fontSize: 32.0)),
+          ),
+          Center(
+            child: Text('セキュリティ', style: TextStyle(fontSize: 32.0)),
+          ),
+          Center(
+            child: Text('AI', style: TextStyle(fontSize: 32.0)),
+          ),
+          Center(
+            child: Text('AI2', style: TextStyle(fontSize: 32.0)),
+          ),
+        ],
+      ),
+    );
+  }
+}
