+++ lib/widgets/custom_tab_bar.dart
@@ -1,29 +1,31 @@
 import 'package:flutter/material.dart';
 
 class CustomTabBar extends StatelessWidget {
+  final List<Tab> tabs;
+  final List<Widget> list;
+  final int initialIndex;
+
   const CustomTabBar({
     Key? key,
-    required this.tab,
+    required this.tabs,
     required this.list,
+    this.initialIndex = 0,
   }) : super(key: key);
-
-  final List<Tab> tab;
-  final List<Widget> list;
 
   @override
   Widget build(BuildContext context) {
     return DefaultTabController(
-      initialIndex: 0, // 最初に表示するタブ
-      length: tab.length,
+      initialIndex: initialIndex,
+      length: tabs.length,
       child: Scaffold(
         appBar: AppBar(
-          bottom: TabBar(
-            isScrollable: true,
-            indicatorColor: Colors.white,
-            labelPadding: const EdgeInsets.symmetric(horizontal: 16.0),
-            tabs: tab,
+          flexibleSpace: Center(
+            child: TabBar(
+              isScrollable: true,
+              labelPadding: const EdgeInsets.symmetric(horizontal: 16.0),
+              tabs: tabs,
+            ),
           ),
-          backgroundColor: Colors.black,
         ),
         body: TabBarView(
           children: list,
